ExampleModel: extract to a separate file
